Update ELIF OPERACII VARHU CHISLA.cpp

+++ ELIF OPERACII VARHU CHISLA.cpp
@@ -78,7 +78,7 @@ int main() {
     else if (nOperator == '/'){
       cout.setf (ios::fixed);
       cout.precision(2);
-      double result = n1 / n2;
+      int result = n1 / n2;
       cout << n1 << " " << nOperator << " " << n2 << " = " << result << endl;
     }
     else if (nOperator == '%') {
